+++ CHANGELOG.md
@@ -6,6 +6,15 @@ Changes:
 
 \-
 
+## v2022-11-04-1
+
+Changes:
+
+- Localized data labels from backend
+- Photo type/quality/min height/max height from backend
+- Parent address
+- "Your previous studies" page qualification checkbox
+
 ## v2022-11-03-2
 
 Changes: